requirements.txt
unidecode~=1.3.8
pillow~=10.3.0